Add basic example of data table with pandas
I added a csv file with some gps data to app/static/data/. I set
up a basic page that reads the data with pandas, generates a table,
then passes the table (as html) to the template file. There are
additional dependencies now, so you'll have to run the `pip install`
step again.